% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bh_rejectC}
\alias{bh_rejectC}
\title{bh_rejectC}
\usage{
bh_rejectC(pvals, alpha, conserv)
}
\arguments{
\item{pvals}{is the vector of pvalues}

\item{alpha}{is the significance level for the BH test}

\item{conserv}{(=TRUE or FALSE). If conserv=TRUE, BHY test is used at level alpha 
otherwise BH test is used at level alpha.}
}
\description{
The rcpp version of bhreject
}
\details{
\code{bh_rejectC} takes a vector of p-values and applies the Benjamini Hochberg (Yekutelli)
multiple testing procedure at significane level alpha, and returns the indices 
of the pvalues which were rejected.
}
