\name{bmdCpp-package}
\alias{bmdCpp-package}
\alias{bmdCpp}
\docType{package}
\title{
\packageTitle{bmdCpp}
}
\description{
\packageDescription{bmdCpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bmdCpp}
\packageIndices{bmdCpp}
This package provides the functions for Bimodule detection.
}
\author{
\packageAuthor{bmdCpp}

Maintainer: \packageMaintainer{bmdCpp}
}
\references{
NA
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}