% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeC}
\alias{initializeC}
\title{inititalizeC}
\usage{
initializeC(n, cors, alpha, conserv)
}
\arguments{
\item{n}{is the sample size}

\item{cors}{is the vector of correlations to the initital node.}

\item{alpha}{is the significance level for the BH test}

\item{conserv}{(=TRUE or FALSE). If conserv=TRUE, BHY test is used at alpha 
otherwise BH test is used at alpha.}
}
\details{
\code{initializeC} takes a vector of correlations, firscher transforms it 
 (to get a z-statistic) and computes the pvalues. The BH (or BHY) procedure is 
 used on these pvales to reject a set of indices which are then returned
}
