% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pvalsC}
\alias{pvalsC}
\title{pvalsC}
\usage{
pvalsC(first, second, f4ColSums, f2, f3, cross)
}
\arguments{
\item{first}{(either X or Y) is the matrix corresponding to the first side.}

\item{second}{(e.g X[,A] or Y[,A]) is the sub-matrix (of the second side)
corresponding to the columns of the A.}

\item{f4ColSums}{is the rowvector columnSums(first^4)}

\item{f2}{is the matrix first^2}

\item{f3}{is the matrix first^3}

\item{cross}{is the matrix cor(first, second)}
}
\description{
The p-values in the one-sided updated step. first corresponds to the side 
(X or Y) the p-values will be generated for. This is side opposite to 
that of the index set A that is being used.
}
\details{
\code{pvalsC} computes the z-statistic corresponding to the sum of 
correlations to the second matrix for each column of the first matrix
(using the variance formula from SteigerHakstian1982). 
The pvalue vector corresponding to the columns of first matrix is returned.
}
