% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateC}
\alias{updateC}
\title{updateC}
\usage{
updateC(first, second, f4ColSums, f2, f3, cross, alpha, conserv)
}
\arguments{
\item{first}{(either X or Y) is the matrix corresponding to the first side.}

\item{second}{(e.g X[,A] or Y[,A]) is the sub-matrix (of the second side)
corresponding to the columns of the A.}

\item{f4ColSums}{is the rowvector columnSums(first^4)}

\item{f2}{is the matrix first^2}

\item{f3}{is the matrix first^3}

\item{cross}{is the matrix cor(first, second)}

\item{alpha}{is the significance level for the BH test}

\item{conserv}{(=TRUE or FALSE). If conserv=TRUE, BHY test is used at alpha 
otherwise BH test is used at alpha.}
}
\description{
The rcpp version of update5. update5 is first step for a one sided update.
One starts with a subset A of X or Y indices, and wishes to get a set B=update5(A)
at the other side. Call the otherside as the first side (or the side being tested)
and side corresponding to A as the second side.
}
\details{
\code{updateC} computes the z-statistic corresponding to the sum of 
correlations to the second matrix for each column of the first matrix
(using the variance formula from SteigerHakstian1982). It converts the 
z-statistic to pvalues and rejects the columns of first matrix which 
were significant. The indicies corresponding to the significant columns are
returned.
}
