\name{bmdCpp-package}
\alias{bmdCpp-package}
\alias{bmdCpp}
\docType{package}
\title{
\packageTitle{bmdCpp}
}
\description{
\packageDescription{bmdCpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bmdCpp}
\packageIndices{bmdCpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bmdCpp}

Maintainer: \packageMaintainer{bmdCpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
